+++ Sources/BFKit/Apple/UIKit/UILabel+Extensions.swift
@@ -4,7 +4,7 @@
 //
 //  The MIT License (MIT)
 //
-//  Copyright (c) 2015 - 2018 Fabrizio Brancati.
+//  Copyright (c) 2015 - 2019 Fabrizio Brancati.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -83,6 +83,7 @@ public extension UILabel {
         guard let text = text else {
             return 0
         }
+        
         return UIFont.calculateHeight(width: frame.size.width, font: font, text: text)
     }