+++ Sources/BFKit/Apple/UIKit/UIViewController+Extensions.swift
@@ -4,7 +4,7 @@
 //
 //  The MIT License (MIT)
 //
-//  Copyright (c) 2015 - 2018 Fabrizio Brancati.
+//  Copyright (c) 2015 - 2019 Fabrizio Brancati.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -39,7 +39,8 @@ public extension UIViewController {
         
         if let coordinator = transitionCoordinator {
             coordinator.animateAlongsideTransition(
-                in: parent?.view, animation: { coordinatorContext in
+                in: parent?.view,
+                animation: { coordinatorContext in
                     selectedIndexPaths.forEach {
                         tableView.deselectRow(at: $0, animated: coordinatorContext.isAnimated)
                     }