+++ Sources/BFKit/Linux/Foundation/Collection+Extensions.swift
@@ -4,7 +4,7 @@
 //
 //  The MIT License (MIT)
 //
-//  Copyright (c) 2015 - 2018 Fabrizio Brancati.
+//  Copyright (c) 2015 - 2019 Fabrizio Brancati.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -42,6 +42,7 @@ public extension Collection {
         guard let json = NSString(data: jsonData, encoding: String.Encoding.utf8.rawValue) else {
             throw BFKitError.jsonSerialization
         }
+        
         return String(describing: json)
     }
 }