+++ Tests/BFKitTests/Apple/WebKit/WKWebViewExtensionTests.swift
@@ -4,7 +4,7 @@
 //
 //  The MIT License (MIT)
 //
-//  Copyright (c) 2015 - 2018 Fabrizio Brancati.
+//  Copyright (c) 2015 - 2019 Fabrizio Brancati.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -39,7 +39,7 @@ internal class WKWebViewExtensionTests: XCTestCase {
   }
   
   internal func testLoadWebsite() {
-    webView.load("https://www.fabriziobrancati.com")
+    webView.load(website: "https://www.fabriziobrancati.com")
     
     webView.evaluateJavaScript("document.getElementsByTagName('html')[0].innerHTML") { html, error in
       if error != nil {
@@ -56,4 +56,3 @@ internal class WKWebViewExtensionTests: XCTestCase {
     }
   }
 }
-